CRAFTERCMS-188 Multiple publishing groups support implemented

+++ alfresco-svcs/src/main/java/org/craftercms/cstudio/impl/service/deployment/PublishingManagerImpl.java
@@ -442,26 +442,27 @@ public class PublishingManagerImpl implements PublishingManager {
 
     @Override
     public void processItem(CopyToEnvironmentItem item) throws DeploymentException {
+        String liveEnvironment = _contentRepository.getLiveEnvironmentName(item.getSite());
+        boolean isLive = false;
+        if (StringUtils.isNotEmpty(liveEnvironment)) {
+            if (liveEnvironment.equals(item.getEnvironment())) {
+                isLive = true;
+            }
+        } else if (LIVE_ENVIRONMENT.equalsIgnoreCase(item.getEnvironment()) || PRODUCTION_ENVIRONMENT.equalsIgnoreCase(item.getEnvironment())) {
+            isLive = true;
+        }
         if (item.getAction() == CopyToEnvironmentItem.Action.DELETE) {
             if (item.getOldPath() != null && item.getOldPath().length() > 0) {
                 _contentRepository.deleteContent(item.getSite(), item.getEnvironment(), item.getOldPath());
                 _contentRepository.clearRenamed(item.getSite(), item.getPath());
             }
             _contentRepository.deleteContent(item.getSite(), item.getEnvironment(), item.getPath());
-            _contentRepository.deleteContent(item);
+            if (isLive) {
+                _contentRepository.deleteContent(item);
+            }
         } else {
             _contentRepository.setSystemProcessing(item.getSite(), item.getPath(), true);
-            String liveEnvironment = _contentRepository.getLiveEnvironmentName(item.getSite());
-            if (StringUtils.isNotEmpty(liveEnvironment)) {
-                if (liveEnvironment.equalsIgnoreCase(item.getEnvironment())) {
-                    if (!_importModeEnabled) {
-                        _contentRepository.createNewVersion(item.getSite(), item.getPath(), item.getSubmissionComment(), true);
-                    } else {
-                        LOGGER.debug("Import mode is ON. Create new version is skipped for [{0}] site \"{1}\"", item.getPath(), item.getSite());
-                    }
-                    _contentRepository.stateTransition(item.getSite(), item.getPath(), TransitionEvent.DEPLOYMENT);
-                }
-            } else if (LIVE_ENVIRONMENT.equalsIgnoreCase(item.getEnvironment()) || PRODUCTION_ENVIRONMENT.equalsIgnoreCase(item.getEnvironment())) {
+            if (isLive) {
                 if (!_importModeEnabled) {
                     _contentRepository.createNewVersion(item.getSite(), item.getPath(), item.getSubmissionComment(), true);
                 } else {
@@ -472,7 +473,9 @@ public class PublishingManagerImpl implements PublishingManager {
             if (item.getAction() == CopyToEnvironmentItem.Action.MOVE) {
                 if (item.getOldPath() != null && item.getOldPath().length() > 0) {
                     _contentRepository.deleteContent(item.getSite(), item.getEnvironment(), item.getOldPath());
-                    _contentRepository.clearRenamed(item.getSite(), item.getPath());
+                    if (isLive) {
+                        _contentRepository.clearRenamed(item.getSite(), item.getPath());
+                    }
                 }
             }
             _contentRepository.copyToEnvironment(item.getSite(), item.getEnvironment(), item.getPath());
@@ -534,25 +537,27 @@ public class PublishingManagerImpl implements PublishingManager {
                     mandatoryDependencies.addAll(processMandatoryDependencies(parentItem, pathsToDeploy, missingDependenciesPaths));
                 }
             }
-        }
-        List<String> dependentPaths = _contentRepository.getDependentPaths(site, path);
-        for (String dependentPath : dependentPaths) {
-            if (_contentRepository.isNew(site, dependentPath) || _contentRepository.isRenamed(site, dependentPath)) {
-                String dependentFullPath = _contentRepository.getFullPath(site, dependentPath);
-                if (!missingDependenciesPaths.contains(dependentFullPath) && !pathsToDeploy.contains(dependentFullPath)) {
-                    try {
-                        _deploymentDAL.cancelWorkflow(site, dependentPath);
-                    } catch (DeploymentDALException e) {
-                        LOGGER.error("Error while canceling workflow for path {0}, site {1}", e, site, dependentPath);
+
+            List<String> dependentPaths = _contentRepository.getDependentPaths(site, path);
+            for (String dependentPath : dependentPaths) {
+                if (_contentRepository.isNew(site, dependentPath) || _contentRepository.isRenamed(site, dependentPath)) {
+                    String dependentFullPath = _contentRepository.getFullPath(site, dependentPath);
+                    if (!missingDependenciesPaths.contains(dependentFullPath) && !pathsToDeploy.contains(dependentFullPath)) {
+                        try {
+                            _deploymentDAL.cancelWorkflow(site, dependentPath);
+                        } catch (DeploymentDALException e) {
+                            LOGGER.error("Error while canceling workflow for path {0}, site {1}", e, site, dependentPath);
+                        }
+                        missingDependenciesPaths.add(dependentFullPath);
+                        CopyToEnvironmentItem dependentItem = createMissingItem(site, dependentPath, item);
+                        processItem(dependentItem);
+                        mandatoryDependencies.add(dependentItem);
+                        mandatoryDependencies.addAll(processMandatoryDependencies(dependentItem, pathsToDeploy, missingDependenciesPaths));
                     }
-                    missingDependenciesPaths.add(dependentFullPath);
-                    CopyToEnvironmentItem dependentItem = createMissingItem(site, dependentPath, item);
-                    processItem(dependentItem);
-                    mandatoryDependencies.add(dependentItem);
-                    mandatoryDependencies.addAll(processMandatoryDependencies(dependentItem, pathsToDeploy, missingDependenciesPaths));
                 }
             }
         }
+
         return mandatoryDependencies;
     }
